--- src/domain/User/index.ts
@@ -1,5 +1,0 @@
-export type User = {
-  email: string // should be VO email
-  token: string // can I type JWT?
-  username: string
-}
