+++ src/modules/article/application/createArticle.spec.ts
@@ -1,4 +1,4 @@
-import { ArticleSlug } from '@domain/Article'
+import { ArticleSlug } from '@@/src/modules/article/domain'
 
 test('just asserting imports work fine in Jest env', () => {
   const slug: ArticleSlug = 'asd'
