+++ src/modules/article/domain/article.types.ts
@@ -4,7 +4,7 @@ export type Article = {
   title: string
   slug: ArticleSlug
   body: string
-  createdAt: Date | string
+  createdAt: Date
   favorited: boolean
 }
 
@@ -12,5 +12,5 @@ export type EditingArticle = {
   title: string
   slug?: ArticleSlug
   body: string
-  createdAt: Date | string
+  createdAt: Date
 }
