+++ src/shared/http/HttpError.spec.ts
