+++ src/shared/http/HttpError.ts
