+++ src/shared/http/HttpResult.ts
@@ -1,4 +1,4 @@
 import { AsyncResult } from '@@/src/shared/Result'
-import { HttpError } from '@@/src/shared/HttpError'
+import { HttpError } from '@@/src/shared/http/HttpError'
 
 export type HttpResult<T> = AsyncResult<T, HttpError | Error>
