+++ src/ui/components/ArticleList.vue
@@ -10,7 +10,7 @@
 
 <script lang="ts">
 import Vue, { PropOptions } from 'vue'
-import { Article } from '@domain/Article'
+import { Article } from '@@/src/modules/article/domain'
 
 export default Vue.extend({
   name: 'ArticleList',
