+++ src/ui/store/article/mutations.ts
@@ -1,4 +1,4 @@
-import { Article } from '@domain/Article'
+import { Article } from '@@/src/modules/article/domain'
 import { MutationTree } from 'vuex'
 import { RootState } from './state'
